requirements.txt
opencv-python
mediapipe
numpy
pillow
pyyaml>=6.0

# dev tools
black
ruff
pytest
